Checkpoint system !
This is the checkpoing system I'm using in my world.
It works, for the most part, so here it is !

The idea is that, once you hit a checkpoint, a teleporter
to the main base appears near the checkpoint. That way, you
can go back to the main area of your world.
Also, hitting the checkpoint moves your spawn point, so that
you can respawn in that place.

Signed-off-by: Voyage <voyage@miouyouyou.fr>